+++ src/sdk/v2/permit.generate.ts
@@ -2,7 +2,7 @@ import {
   EIP712Message,
   EIP712Types,
 } from "../../extensions/access_control/EIP712";
-import { PermissionV2 } from "../types";
+import { PermissionV2 } from "./types";
 import { ZeroAddress } from "ethers";
 
 const PermitV2SignatureAllFields = [
